docs(slate-instagram): add comments and README
