+++ packages/slate-instagram/src/controller.ts
@@ -1,10 +1,13 @@
-import { Block, Editor } from 'slate';
+import { Editor } from 'slate';
 import { NodeType } from '@artibox/slate-common';
 import { PARAGRAPH_TYPE } from '@artibox/slate-common/constants/paragraph';
 import { getInstagramUrlFromEmbedCode } from './utils/get-instagram-url-from-embed-code';
+import { createInstagramBlock } from './utils/create-instagram-block';
 
 export interface InstagramController {
-  createBlock(url: string): Block;
+  /**
+   * Add the instagram block to editor.
+   */
   add(editor: Editor, embedCode: string): Editor;
 }
 
@@ -12,7 +15,6 @@ export type CreateInstagramControllerConfig = NodeType;
 
 export function createInstagramController(config: CreateInstagramControllerConfig): InstagramController {
   const { type } = config;
-  const createBlock: InstagramController['createBlock'] = url => Block.fromJSON({ type, data: { url } });
   const add: InstagramController['add'] = (editor, embedCode) => {
     const url = getInstagramUrlFromEmbedCode(embedCode);
 
@@ -20,8 +22,8 @@ export function createInstagramController(config: CreateInstagramControllerConfi
       return editor;
     }
 
-    return editor.insertBlock(createBlock(url)).insertBlock(PARAGRAPH_TYPE);
+    return editor.insertBlock(createInstagramBlock(type, url)).insertBlock(PARAGRAPH_TYPE);
   };
 
-  return { createBlock, add };
+  return { add };
 }
