+++ packages/slate-instagram/src/instagram.ts
@@ -9,7 +9,10 @@ import { createInstagramSchema } from './schema';
 export type InstagramForPluginConfig = Omit<CreateInstagramRendererConfig, 'type'>;
 
 export interface InstagramForToolHookConfig {
-  setToolInput?: (editor: Editor, toolInput: InputConfig) => Editor | void;
+  /**
+   * The callback method override the origin setInputConfig provided by toolbar.
+   */
+  setInputConfig?: (editor: Editor, inputConfig: InputConfig) => Editor | void;
 }
 
 export type Instagram = NodeType &
@@ -32,21 +35,21 @@ export function createInstagram(config?: CreateInstagramConfig): Instagram {
         schema: createInstagramSchema({ type })
       };
     },
-    forToolHook(config: InstagramForToolHookConfig): ToolHook {
-      const { setToolInput } = config;
-      const toolInput: InputConfig = {
+    forToolHook(config): ToolHook {
+      const { setInputConfig } = config || {};
+      const inputConfig: InputConfig = {
         getPlaceholder: locale => locale.editor.instagram.inputPlaceholder,
         onConfirm: controller.add
       };
 
-      return (editor, defaultSetToolInput) => ({
+      return (editor, defaultSetInputConfig) => ({
         onMouseDown: useCallback(() => {
-          if (setToolInput) {
-            setToolInput(editor, toolInput);
+          if (setInputConfig) {
+            setInputConfig(editor, inputConfig);
           } else {
-            defaultSetToolInput(toolInput);
+            defaultSetInputConfig(inputConfig);
           }
-        }, [editor, defaultSetToolInput])
+        }, [editor, defaultSetInputConfig])
       });
     }
   };
