+++ packages/slate-instagram/src/utils/create-instagram-block.ts
@@ -1,0 +1,5 @@
+import { Block } from 'slate';
+
+export function createInstagramBlock(type: string, url: string) {
+  return Block.fromJSON({ type, data: { url } });
+}
