+++ packages/slate-instagram/src/utils/get-instagram-props-from-block.ts
@@ -1,5 +1,5 @@
 import { Block } from 'slate';
-import { InstagramProps } from '../types';
+import { InstagramProps } from '../typings';
 
 export function getInstagramPropsFromBlock(block: Block): InstagramProps {
   return { url: block.data.get('url') || '' };
