+++ stories/rich-editor/plugins.ts
@@ -85,7 +85,7 @@ export const plugins: Plugin[] = [
       TOOLBAR_DIVIDER,
       { icon: SeparationLine, hook: separationLine.forToolHook() },
       { icon: Video, hook: video.forToolHook({ setToolInput: inputBlock.start }) },
-      { icon: Instagram, hook: instagram.forToolHook({ setToolInput: inputBlock.start }) },
+      { icon: Instagram, hook: instagram.forToolHook({ setInputConfig: inputBlock.start }) },
       { icon: Facebook, hook: facebook.forToolHook({ setInputConfig: inputBlock.start }) }
     ]
   })
